(BOLT-255) Move high level logic to SSH transport
This migrates much of the logic from the {SSH,WinRM}::Connection objects
into the transports directly. We now only use the Connection for
primitive operations like creating files and executing commands, while
the logic of how to upload, stage and run tasks, etc is handled by the
transport.

Effectively, the connection is responsible for most of the work that
requires knowledge of the target, and the transport has logic that
operates irrespective of the target.
