+++ lib/bolt/transport/winrm/connection.rb
@@ -18,9 +18,6 @@ module Bolt
           @logger = Logging.logger[@target.host]
         end
 
-        STDIN_METHODS       = %w[both stdin].freeze
-        ENVIRONMENT_METHODS = %w[both environment].freeze
-
         HTTP_PORT = 5985
         HTTPS_PORT = 5986
 
@@ -394,7 +391,6 @@ PS
           end
           result_output
         end
-        private :execute
 
         # 10 minutes in seconds
         DEFAULT_EXECUTION_TIMEOUT = 10 * 60
@@ -428,41 +424,6 @@ PS
         PS_ARGS = %w[
           -NoProfile -NonInteractive -NoLogo -ExecutionPolicy Bypass
         ].freeze
-
-        def powershell_file?(path)
-          Pathname(path).extname.casecmp('.ps1').zero?
-        end
-
-        def process_from_extension(path)
-          case Pathname(path).extname.downcase
-          when '.rb'
-            [
-              'ruby.exe',
-              ['-S', "\"#{path}\""]
-            ]
-          when '.ps1'
-            [
-              'powershell.exe',
-              [*PS_ARGS, '-File', "\"#{path}\""]
-            ]
-          when '.pp'
-            [
-              'puppet.bat',
-              ['apply', "\"#{path}\""]
-            ]
-          else
-            # Run the script via cmd, letting Windows extension handling determine how
-            [
-              'cmd.exe',
-              ['/c', "\"#{path}\""]
-            ]
-          end
-        end
-
-        def upload(source, destination, _options = nil)
-          write_remote_file(source, destination)
-          Bolt::Result.for_upload(target, source, destination)
-        end
 
         def write_remote_file(source, destination)
           fs = ::WinRM::FS::FileManager.new(@connection)
@@ -503,93 +464,6 @@ PS
 Remove-Item -Force "#{dest}"
 Remove-Item -Force "#{dir}"
 PS
-          end
-        end
-
-        def run_command(command, _options = nil)
-          output = execute(command)
-          Bolt::Result.for_command(target, output.stdout.string, output.stderr.string, output.exit_code)
-        end
-
-        def run_script(script, arguments, _options = nil)
-          with_remote_file(script) do |remote_path|
-            if powershell_file?(remote_path)
-              mapped_args = arguments.map do |a|
-                "$invokeArgs.ArgumentList += @'\n#{a}\n'@"
-              end.join("\n")
-              output = execute(<<-PS)
-$invokeArgs = @{
-  ScriptBlock = (Get-Command "#{remote_path}").ScriptBlock
-  ArgumentList = @()
-}
-#{mapped_args}
-
-try
-{
-  Invoke-Command @invokeArgs
-}
-catch
-{
-  exit 1
-}
-          PS
-            else
-              path, args = *process_from_extension(remote_path)
-              args += escape_arguments(arguments)
-              output = execute_process(path, args)
-            end
-            Bolt::Result.for_command(target, output.stdout.string, output.stderr.string, output.exit_code)
-          end
-        end
-
-        def run_task(task, input_method, arguments, _options = nil)
-          if STDIN_METHODS.include?(input_method)
-            stdin = JSON.dump(arguments)
-          end
-
-          if ENVIRONMENT_METHODS.include?(input_method)
-            arguments.each do |(arg, val)|
-              cmd = "[Environment]::SetEnvironmentVariable('PT_#{arg}', '#{val}')"
-              result = execute(cmd)
-              if result.exit_code != 0
-                raise EnvironmentVarError(var, value)
-              end
-            end
-          end
-
-          with_remote_file(task) do |remote_path|
-            output =
-              if powershell_file?(remote_path) && stdin.nil?
-                # NOTE: cannot redirect STDIN to a .ps1 script inside of PowerShell
-                # must create new powershell.exe process like other interpreters
-                # fortunately, using PS with stdin input_method should never happen
-                if input_method == 'powershell'
-                  execute(<<-PS)
-$private:taskArgs = Get-ContentAsJson (
-  $utf8.GetString([System.Convert]::FromBase64String('#{Base64.encode64(JSON.dump(arguments))}'))
-)
-try { & "#{remote_path}" @taskArgs } catch { exit 1 }
-              PS
-                else
-                  execute(%(try { & "#{remote_path}" } catch { exit 1 }))
-                end
-              else
-                path, args = *process_from_extension(remote_path)
-                execute_process(path, args, stdin)
-              end
-            Bolt::Result.for_task(target, output.stdout.string,
-                                  output.stderr.string,
-                                  output.exit_code)
-          end
-        end
-
-        def escape_arguments(arguments)
-          arguments.map do |arg|
-            if arg =~ / /
-              "\"#{arg}\""
-            else
-              arg
-            end
           end
         end
       end
